# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service Provider",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/public/api",
            "description": "main server"
        },
        {
            "url": "/backend/public/api",
            "description": "local server"
        }
    ],
    "paths": {
        "/agent/register": {
            "post": {
                "tags": [
                    "Auth-Agent_register"
                ],
                "summary": "Register as a new agent",
                "operationId": "authRegister",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Registration model",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterAdminModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "RegisterResult response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResultRegisterResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/register-user": {
            "post": {
                "tags": [
                    "Auth-client_register"
                ],
                "summary": "Register as a new user",
                "operationId": "authRegister",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Registration model",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "RegisterResult response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResultRegisterResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "Auth-login"
                ],
                "summary": "Login as a Agent and User",
                "operationId": "authLogin",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Login model",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginAdminPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Profile response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginAdminResultApi"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/verify-account": {
            "post": {
                "tags": [
                    "Auth-login"
                ],
                "summary": "Verify account with changing its status to active",
                "operationId": "4b65d04fa8c4ba8ede59f79cca859ea0",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "ConfirmAccountModel model",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfirmAccountModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Profile response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResultApi"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/change-password": {
            "post": {
                "tags": [
                    "Auth-login"
                ],
                "summary": "Change Password For Agent And User",
                "operationId": "authChangePhone",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "ChangePhoneNumberKernelModel model",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordAgentPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/forget-password": {
            "post": {
                "tags": [
                    "Auth-login"
                ],
                "summary": "Resend the code if not received",
                "operationId": "524f542eb7d698bf7911f3a990e67711",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "phone of admin",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/reset-password": {
            "post": {
                "tags": [
                    "Auth-login"
                ],
                "summary": "Verify account with changing its status to active",
                "operationId": "f4b93abe70b831f891336181f58e2457",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "ConfirmAccountModel model",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgetPasswordConfirm"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Profile response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResultApi"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "tags": [
                    "Auth-login"
                ],
                "summary": "Logout from system",
                "operationId": "d061a135544d44fb15e37894c26bb694",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/delete-account": {
            "post": {
                "tags": [
                    "Auth-login"
                ],
                "summary": "Delete as a user and agent",
                "operationId": "31e90a52efbe5814cd08c4fd60f4f6f4",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "LoginPayload model",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/check-token": {
            "get": {
                "tags": [
                    "Auth-login"
                ],
                "summary": "Test a token for expiration",
                "operationId": "8a991b2d8863c5aa81b26ae303737e96",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResultRegisterResult": {
                "title": "ApiResultRegisterResult model",
                "description": "ApiResultRegisterResult model",
                "properties": {
                    "result": {
                        "$ref": "#/components/schemas/RegisterResult"
                    },
                    "isOk": {
                        "title": "isOk",
                        "description": "Indicates if the response is ok or not",
                        "type": "boolean"
                    },
                    "message": {
                        "$ref": "#/components/schemas/ApiMessage"
                    }
                },
                "type": "object"
            },
...
            "ConfirmAccountModel": {
                "title": "ConfirmAccountModel model",
                "description": "ConfirmAccountModel model",
                "properties": {
                    "phone": {
                        "title": "phone",
                        "description": "The phone of user to verify",
                        "type": "string"
                    },
                    "code": {
                        "title": "code",
                        "description": "Verification Code",
                        "type": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "RegisterModel": {
                "title": "RegisterModel model",
                "description": "RegisterModel model",
                "properties": {
                    "fullname": {
                        "title": "fullname",
                        "description": "fullname",
                        "type": "string"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "Phone",
                        "type": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "User  password",
                        "type": "string"
                    },
                    "ConfirmPassword": {
                        "title": "confirm_password",
                        "description": "User  password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "LoginAdminPayload": {
                "title": "LoginAdminPayload model",
                "description": "LoginAdminPayload model",
                "properties": {
                    "phone": {
                        "title": "phone",
                        "description": "Phone number",
                        "type": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "Account Password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "LoginAdminResultApi": {
                "title": "LoginAdminResultApi model",
                "description": "LoginAdminResultApi model",
                "properties": {
                    "result": {
                        "$ref": "#/components/schemas/LoginAdminResult"
                    },
                    "isOk": {
                        "title": "isOk",
                        "description": "Indicates if the response is ok or not",
                        "type": "boolean"
                    },
                    "message": {
                        "$ref": "#/components/schemas/ApiMessage"
                    }
                },
                "type": "object"
            },
...
            "RegisterAdminModel": {
                "title": "RegisterAdminModel model",
                "description": "RegisterAdminModel model",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string"
                    },
                    "job_type_id": {
                        "title": "job_type_id",
                        "description": "job_type_id",
                        "type": "integer"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "Phone",
                        "type": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "Account Password",
                        "type": "string"
                    },
                    "country_id": {
                        "title": "country_id",
                        "description": "country_id",
                        "type": "integer"
                    },
                    "city_id": {
                        "title": "city_id",
                        "description": "city_id",
                        "type": "integer"
                    },
                    "address": {
                        "title": "address",
                        "description": "address",
                        "type": "string"
                    },
                    "linephone": {
                        "title": "linephone",
                        "description": "linephone",
                        "type": "string"
                    },
                    "description": {
                        "title": "description",
                        "description": "description",
                        "type": "string"
                    },
                    "contact_media_ids": {
                        "title": "contact_media_ids",
                        "description": "contact_media_ids",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "sections_ids": {
                        "title": "sections_ids",
                        "description": "sections_ids",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "image_license[]": {
                        "description": "Item image_license",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary"
                        }
                    },
                    "image_commercial[]": {
                        "description": "Item image_commercial",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Authentication Bearer Token",
                "name": "Authentication Bearer Token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}